+++ frontend/src/domains/admin/pages/agenda/chat/index.tsx
@@ -1,10 +1,10 @@
-import { useLayoutEffect, useRef, useState } from 'react';
-import ChatPage, { ChatRoomInfo } from '../../chat-page';
+import { useEffect, useLayoutEffect, useRef, useState } from 'react';
+import ChatPage, { ChatRoomInfo } from './chat-page';
 import { useParams, useSearchParams } from 'react-router-dom';
 import { formatChatData } from '@/utils/chat/formatChatData.ts';
 import api from '@/utils/api.ts';
 import useInfiniteScroll from '@/hooks/useInfiniteScroll';
-import { ChatData } from '../../chat-page/ChatData';
+import { ChatData } from './chat-page/ChatData';
 import { useScroll } from '@/hooks/useScrollBottom';
 
 const AgendaChatPage = () => {
@@ -138,7 +138,6 @@ const AgendaChatPage = () => {
         setTriggerDownItem(lastItemRef);
       }}
       onMessageSend={(data: ChatData) => {
-        // console.log('gethasmore', getHasDownMore());
         if (!getHasDownMore()) {
           isLive.current = true;
           setChatData((prev) => [...prev, data]);
